requirements.txt
beautifulsoup4==4.12.2
gensim==4.3.1
huggingface-hub==0.13.3
matplotlib==3.7.1
networkx==3.1
nltk==3.8.1
numpy==1.24.2
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.0
Pillow==9.5.0
requests==2.28.2
scikit-learn==1.2.2
scikit-learn-extra==0.3.0
scikit-surprise==1.1.3
scipy==1.10.1
seaborn==0.12.2
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
wordcloud==1.9.2
contractions==0.1.73
